Implement Backend Models and Database Setup
This commit introduces several changes to the backend of the application. The main focus is on setting up the database and implementing the necessary models.

The following models have been added:
- BaseModel: This is the base class from which all other models will inherit. It includes common attributes like id, created_at, and updated_at.
- Class: This model represents a gym class. It includes attributes like name, description, instructor_id, location, date_and_time, duration_minutes, capacity, and image_url.
- Instructor: This model represents an instructor. It includes attributes like email, password, first_name, last_name, picture, phone, bio, approach, specializations, experience_years, and classes.

The DBStorage class has been implemented to interact with the MySQL database. It includes methods for querying, adding, saving, deleting, and reloading data from the database.

A RedisEngine class has also been implemented to interact with a Redis server. It includes methods for checking server connection, getting, setting, and deleting keys.

The .gitignore file has been updated to ignore certain files and directories related to the backend and the client.

A setup_mysql.sql file has been added to automate the process of setting up the MySQL database.

The backend/app.py file has been updated to import the dbstorage from models and to close the dbstorage after each request.

The alembic.ini file has been added for database migration purposes.

Tests have been added to test the functionality of the new models and database setup.