Refactor backend and add user model
This commit includes several changes to the backend of the application. The main change is the addition of a User model, which includes a relationship to the Class model through a secondary table. This allows users to be associated with multiple classes.

The Instructor model has also been updated to include a gender field and the name field has been consolidated from first_name and last_name to just name.

The DBStorage class has been updated to include the User model and a new method, get_by, has been added to allow retrieval of objects based on a key-value pair.

The app.py file has been updated to use the renamed storage module instead of dbstorage.

The requirements.txt file has been updated to include new dependencies and a start-server.sh script has been added to automate the process of starting the server.

Finally, new authentication views have been added for both users and instructors, allowing them to login and register.